+++ common/persistence/sql/storage/sqlshared/events.go
@@ -18,33 +18,12 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package mysql
+package sqlshared
 
 import (
 	"database/sql"
 
 	"github.com/uber/cadence/common/persistence/sql/storage/sqldb"
-)
-
-const (
-	// below are templates for history_node table
-	addHistoryNodesQry = `INSERT INTO history_node (` +
-		`shard_id, tree_id, branch_id, node_id, txn_id, data, data_encoding) ` +
-		`VALUES (:shard_id, :tree_id, :branch_id, :node_id, :txn_id, :data, :data_encoding) `
-
-	getHistoryNodesQry = `SELECT node_id, txn_id, data, data_encoding FROM history_node ` +
-		`WHERE shard_id = ? AND tree_id = ? AND branch_id = ? AND node_id >= ? and node_id < ? ORDER BY shard_id, tree_id, branch_id, node_id, txn_id LIMIT ? `
-
-	deleteHistoryNodesQry = `DELETE FROM history_node WHERE shard_id = ? AND tree_id = ? AND branch_id = ? AND node_id >= ? `
-
-	// below are templates for history_tree table
-	addHistoryTreeQry = `INSERT INTO history_tree (` +
-		`shard_id, tree_id, branch_id, data, data_encoding) ` +
-		`VALUES (:shard_id, :tree_id, :branch_id, :data, :data_encoding) `
-
-	getHistoryTreeQry = `SELECT branch_id, data, data_encoding FROM history_tree WHERE shard_id = ? AND tree_id = ? `
-
-	deleteHistoryTreeQry = `DELETE FROM history_tree WHERE shard_id = ? AND tree_id = ? AND branch_id = ? `
 )
 
 // For history_node table:
@@ -53,13 +32,13 @@ const (
 func (mdb *DB) InsertIntoHistoryNode(row *sqldb.HistoryNodeRow) (sql.Result, error) {
 	// NOTE: MySQL 5.6 doesn't support clustering order, to workaround, we let txn_id multiple by -1
 	*row.TxnID *= -1
-	return mdb.conn.NamedExec(addHistoryNodesQry, row)
+	return mdb.conn.NamedExec(mdb.driver.AddHistoryNodesQry(), row)
 }
 
 // SelectFromHistoryNode reads one or more rows from history_node table
 func (mdb *DB) SelectFromHistoryNode(filter *sqldb.HistoryNodeFilter) ([]sqldb.HistoryNodeRow, error) {
 	var rows []sqldb.HistoryNodeRow
-	err := mdb.conn.Select(&rows, getHistoryNodesQry,
+	err := mdb.conn.Select(&rows, mdb.driver.GetHistoryNodesQry(),
 		filter.ShardID, filter.TreeID, filter.BranchID, *filter.MinNodeID, *filter.MaxNodeID, *filter.PageSize)
 	// NOTE: since we let txn_id multiple by -1 when inserting, we have to revert it back here
 	for _, row := range rows {
@@ -70,24 +49,24 @@ func (mdb *DB) SelectFromHistoryNode(filter *sqldb.HistoryNodeFilter) ([]sqldb.H
 
 // DeleteFromHistoryNode deletes one or more rows from history_node table
 func (mdb *DB) DeleteFromHistoryNode(filter *sqldb.HistoryNodeFilter) (sql.Result, error) {
-	return mdb.conn.Exec(deleteHistoryNodesQry, filter.ShardID, filter.TreeID, filter.BranchID, *filter.MinNodeID)
+	return mdb.conn.Exec(mdb.driver.DeleteHistoryNodesQry(), filter.ShardID, filter.TreeID, filter.BranchID, *filter.MinNodeID)
 }
 
 // For history_tree table:
 
 // InsertIntoHistoryTree inserts a row into history_tree table
 func (mdb *DB) InsertIntoHistoryTree(row *sqldb.HistoryTreeRow) (sql.Result, error) {
-	return mdb.conn.NamedExec(addHistoryTreeQry, row)
+	return mdb.conn.NamedExec(mdb.driver.AddHistoryTreeQry(), row)
 }
 
 // SelectFromHistoryTree reads one or more rows from history_tree table
 func (mdb *DB) SelectFromHistoryTree(filter *sqldb.HistoryTreeFilter) ([]sqldb.HistoryTreeRow, error) {
 	var rows []sqldb.HistoryTreeRow
-	err := mdb.conn.Select(&rows, getHistoryTreeQry, filter.ShardID, filter.TreeID)
+	err := mdb.conn.Select(&rows, mdb.driver.GetHistoryTreeQry(), filter.ShardID, filter.TreeID)
 	return rows, err
 }
 
 // DeleteFromHistoryTree deletes one or more rows from history_tree table
 func (mdb *DB) DeleteFromHistoryTree(filter *sqldb.HistoryTreeFilter) (sql.Result, error) {
-	return mdb.conn.Exec(deleteHistoryTreeQry, filter.ShardID, filter.TreeID, *filter.BranchID)
+	return mdb.conn.Exec(mdb.driver.DeleteHistoryTreeQry(), filter.ShardID, filter.TreeID, *filter.BranchID)
 }
