+++ common/persistence/sql/storage/sqlshared/queue.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package mysql
+package sqlshared
 
 import (
 	"database/sql"
